refactor: `core` 모듈을 `api` 및 `domain`으로 분리
