+++ core/api/src/main/resources/schema.sql
