+++ core/domain/src/main/java/com/nilgil/book/core/domain/BookRepository.java
@@ -1,0 +1,9 @@
+package com.nilgil.book.core.domain;
+
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.Optional;
+
+public interface BookRepository extends CrudRepository<BookRow, String> {
+    Optional<BookRow> findByIsbn13(String isbn13);
+}
