+++ core/domain/src/main/java/com/nilgil/book/core/domain/BookRow.java
@@ -1,0 +1,20 @@
+package com.nilgil.book.core.domain;
+
+import org.springframework.data.annotation.Id;
+import org.springframework.data.relational.core.mapping.Table;
+
+@Table("book_read_model")
+public record BookRow(
+        @Id
+        String isbn13,
+        String title,
+        String subtitle,
+        String description,
+        String image,
+        String author,
+        String translator,
+        String publisher,
+        String publishedDate,
+        String pageCount
+) {
+}
