+++ core/domain/src/main/java/com/nilgil/book/core/domain/InvalidIsbnException.java
@@ -1,0 +1,11 @@
+package com.nilgil.book.core.domain;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(code = HttpStatus.BAD_REQUEST)
+public class InvalidIsbnException extends RuntimeException {
+    public InvalidIsbnException(String message) {
+        super(message);
+    }
+}
