--- core/src/main/java/com/nilgil/book/core/query/BookReadService.java
@@ -1,17 +1,0 @@
-package com.nilgil.book.core.query;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class BookReadService {
-
-    private final BookRepository bookRepository;
-
-    public BookDetailResponse getDetailByIsbn(Isbn isbn) {
-        return bookRepository.findByIsbn13(isbn.asIsbn13())
-                .map(BookDetailResponse::from)
-                .orElseThrow(() -> new BookNotFoundException(isbn));
-    }
-}
